move api config structs from daemon to api/types
 - ContainerLogsConfig
 - ContainerStatsConfig
 - ContainersConfig
 - ContainerAttachWithLogsConfig
 - ContainerWsAttachWithLogsConfig
